refactor: rename module `graph` -> `graph_util`
